Rename many modules and change access with __init__.py files
