+++ ironflow/__init__.py
@@ -1,1 +1,8 @@
-from ironflow.ironflow.gui import GUI
+# coding: utf-8
+# Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
+# Distributed under the terms of "New BSD License", see the LICENSE file.
+"""
+Combines ryven, ipywidgets/ipycanvas, and pyiron to facilitate visual scripting of pyiron workflows.
+"""
+
+from ironflow.gui.gui import GUI
