+++ ironflow/gui/__init__.py
@@ -4,7 +4,3 @@
 """
 Code for interfacing with ryven (and pyrion-specific ryven nodes) via ipywidgets and ipython.
 """
-
-from ironflow.ironflow.gui import GUI
-from ironflow.main.node import Node
-from ironflow.main import NodeInputBP, NodeOutputBP, dtypes
