+++ ironflow/gui/boxes/__init__.py
@@ -1,0 +1,12 @@
+# coding: utf-8
+# Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
+# Distributed under the terms of "New BSD License", see the LICENSE file.
+"""
+Component widgets from ipywidgets that get assembled to make the top-level gui.
+"""
+
+from ironflow.gui.boxes.node_interface import NodeController, NodePresenter
+from ironflow.gui.boxes.text_output import TextOut
+from ironflow.gui.boxes.user_input import UserInput
+from ironflow.gui.boxes.toolbar import Toolbar
+from ironflow.gui.boxes.flow import FlowBox
