+++ ironflow/gui/boxes/node_interface/base.py
@@ -11,7 +11,7 @@ from abc import ABC, abstractmethod
 
 import ipywidgets as widgets
 
-from ironflow.ironflow.boxes.base import Box
+from ironflow.gui.boxes.base import Box
 
 
 class NodeInterfaceBase(Box, ABC):
