+++ ironflow/gui/boxes/text_output.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 import ipywidgets as widgets
 from IPython.display import display
 
-from ironflow.ironflow.boxes.base import Box
+from ironflow.gui.boxes.base import Box
 
 
 class TextOut(Box):
