+++ ironflow/gui/canvas_widgets/buttons.py
@@ -10,14 +10,14 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Optional
 
-from ryvencore.NodePort import NodePort, NodeInput, NodeOutput
-
-from ironflow.ironflow.canvas_widgets.base import CanvasWidget, HideableWidget
-from ironflow.ironflow.canvas_widgets.layouts import ButtonLayout
+from ironflow.gui.canvas_widgets.base import CanvasWidget, HideableWidget
+from ironflow.gui.canvas_widgets.layouts import ButtonLayout
+from ironflow.model import NodeInput, NodeOutput
 
 if TYPE_CHECKING:
-    from ironflow.ironflow.canvas_widgets.base import Number
-    from ironflow.ironflow.canvas_widgets.nodes import NodeWidget
+    from ironflow.gui.canvas_widgets.base import Number
+    from ironflow.gui.canvas_widgets.nodes import NodeWidget
+    from ironflow.model import NodePort
 
 
 class ButtonWidget(CanvasWidget, ABC):
