+++ ironflow/gui/canvas_widgets/ports.py
@@ -9,13 +9,13 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Optional
 
-from ironflow.ironflow.canvas_widgets.base import HideableWidget, CanvasWidget
-from ironflow.ironflow.canvas_widgets.layouts import PortLayout
+from ironflow.gui.canvas_widgets.base import HideableWidget, CanvasWidget
+from ironflow.gui.canvas_widgets.layouts import PortLayout
 
 if TYPE_CHECKING:
-    from ironflow.ironflow.canvas_widgets.flow import FlowCanvas
-    from ironflow.ironflow.canvas_widgets.base import Number
-    from ryvencore.NodePort import NodePort
+    from ironflow.gui.canvas_widgets.flow import FlowCanvas
+    from ironflow.gui.canvas_widgets.base import Number
+    from ironflow.model import NodePort
 
 
 class PortWidget(HideableWidget):
