--- ironflow/ironflow/boxes/__init__.py
@@ -1,12 +1,0 @@
-# coding: utf-8
-# Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
-# Distributed under the terms of "New BSD License", see the LICENSE file.
-"""
-Component widgets from ipywidgets that get assembled to make the top-level gui.
-"""
-
-from ironflow.ironflow.boxes.node_interface import NodeController, NodePresenter
-from ironflow.ironflow.boxes.text_output import TextOut
-from ironflow.ironflow.boxes.user_input import UserInput
-from ironflow.ironflow.boxes.toolbar import Toolbar
-from ironflow.ironflow.boxes.flow import FlowBox
