+++ ironflow/model/__init__.py
@@ -2,8 +2,13 @@
 # Copyright (c) Max-Planck-Institut für Eisenforschung GmbH - Computational Materials Design (CM) Department
 # Distributed under the terms of "New BSD License", see the LICENSE file.
 """
-More detailed representations of data from the nodes.
+Wrappers and extensions of the underlying ryven model.
 """
 
-from ironflow.ironflow.boxes.node_interface.control import NodeController
-from ironflow.ironflow.boxes.node_interface.representation import NodePresenter
+from ryvencore import (
+    dtypes,
+    Flow,
+    NodeInputBP,
+    NodeOutputBP,
+)
+from ryvencore.NodePort import NodePort, NodeInput, NodeOutput
