+++ ironflow/nodes/pyiron/atomistics_nodes.py
@@ -18,9 +18,9 @@ from pyiron_atomistics import Project
 from pyiron_atomistics.atomistics.job.atomistic import AtomisticGenericJob
 from pyiron_atomistics.lammps import list_potentials
 
-from ironflow.main import NodeInputBP, NodeOutputBP, dtypes
-from ironflow.main.node import Node
-from ironflow.ironflow.canvas_widgets.nodes import ButtonNodeWidget
+from ironflow.gui.canvas_widgets import ButtonNodeWidget
+from ironflow.model import dtypes, NodeInputBP, NodeOutputBP
+from ironflow.model.node import Node
 from ironflow.nodes.std.special_nodes import DualNodeBase
 
 if TYPE_CHECKING:
