+++ ironflow/nodes/std/special_nodes.py
@@ -1,9 +1,9 @@
 import code
 from contextlib import redirect_stdout, redirect_stderr
 
-from ironflow.main import NodeInputBP, NodeOutputBP, dtypes
-from ironflow.main.node import Node, PlaceholderWidgetsContainer
-from ironflow.ironflow.boxes.node_interface.input_widgets import SliderControl
+from ironflow.model import dtypes, NodeInputBP, NodeOutputBP
+from ironflow.model.node import Node, PlaceholderWidgetsContainer
+from ironflow.gui.boxes.node_interface.input_widgets import SliderControl
 
 widgets = PlaceholderWidgetsContainer()
